BitcoindRegtestSetup: Move RPC calls to RemoteBitcoind
